fix: fifth review - MATCH/XLOOKUP semantics, int cast guards, TRIM perf
- MATCH matchType=-1: return first element >= lookupValue (was incorrectly scanning to last)
- XLOOKUP matchMode=1: return first element >= lookupValue (same logic fix)
- ROUND/ROUNDDOWN/ROUNDUP/TRUNC: double.IsFinite guard before (int) cast of digits arg; ROUND also range-checks [-15,15]
- COMBIN/PERMUT: double.IsFinite + range check before (int) casts
- ADDRESS: reject negative/zero row or column (would wrap uint to ~4B and produce garbage output)
- TEXTJOIN: propagate ErrorValue for args[0] and args[1] before ToBool call
- SUBSTITUTE: validate instanceNum >= 1, return #VALUE! for 0 or negative
- TRIM: replace O(n²) while-loop with compiled Regex for multi-space collapse
- GridRange.CellCount: change return type from uint to long (uint overflows for ranges > ~4B cells)

+++ Freexcel/src/Freexcel.Core.Model/CellAddress.cs
@@ -116,7 +116,7 @@ public readonly record struct GridRange
     public uint ColCount => End.Col - Start.Col + 1;
 
     /// <summary>Total number of cells in this range.</summary>
-    public uint CellCount => RowCount * ColCount;
+    public long CellCount => (long)RowCount * ColCount;
 
     /// <summary>
     /// Enumerate all cell addresses in this range, row by row.
